+++ index.js
@@ -57,6 +57,21 @@ protractorUtil.logDebug = function() {};
 protractorUtil.logInfo = console.info;
 protractorUtil.logError = console.error;
 
+protractorUtil.takeDump = function(context, done) {
+  if (context.config.dump) {
+    try {
+      var result = context.config.dump(function(err, result1) {
+        done(result1);
+      });
+      if (result) {
+        return done(result);
+      }
+    } catch (err) {
+      protractorUtil.logError('Unable to execute spec\'s dump', err);
+      done();
+    }
+  }
+}
 protractorUtil.forEachBrowser = function(action) {
   function catchError(err) {
     console.warn('Unknown error:');
@@ -123,6 +138,29 @@ protractorUtil.takeLogs = function(context, report) {
   protractorUtil.forEachBrowser(takeLog);
 };
 
+protractorUtil.takeRawHtml = function(context, report) {
+  function takeInstanceRawHtml(browserInstance, browserName, cb) {
+    var snapshotFile = 'htmls/' + uuid.v1() + '.html';
+    // protractorUtil.logDebug('Taking raw HTML ' + snapshotFile + ' from browser instance ' + browserName);
+    var finalFile = context.config.screenshotPath + '/' + snapshotFile;
+
+    browserInstance.getPageSource().then(function(html) {
+        fse.writeFile(finalFile, html, 'utf8', function(err) {
+          if (err) {
+            return cb(err);
+          }
+          report(snapshotFile, browserName, finalFile, browserInstance, cb);
+        });
+      },
+      function(err) {
+        console.warn('Error in browser instance ' + browserName + ' while taking the raw html: ' + finalFile + ' - ' + err.message);
+        cb(err);
+      });
+  }
+
+  protractorUtil.forEachBrowser(takeInstanceRawHtml);
+};
+
 protractorUtil.takeScreenshotOnExpectDone = function(context) {
   //Takes screen shot for expect failures
   var originalAddExpectationResult = jasmine.Spec.prototype.addExpectationResult;
@@ -133,6 +171,7 @@ protractorUtil.takeScreenshotOnExpectDone = function(context) {
 
     expectation.screenshots = [];
     expectation.logs = [];
+    expectation.htmls = [];
     expectation.when = now.toDate();
 
     if (!passed && context.config.pauseOn === 'failure') {
@@ -144,18 +183,32 @@ protractorUtil.takeScreenshotOnExpectDone = function(context) {
 
     var makeScreenshotsFromEachBrowsers = false;
     var makeAsciiLog = false;
+    var makeHtmlSnapshot = false;
+    var makeDump = false;
     if (protractorUtil.test) {
       if (passed) {
         protractorUtil.test.passedExpectations.push(expectation);
         makeScreenshotsFromEachBrowsers = context.config.screenshotOnExpect === 'failure+success';
         makeAsciiLog = context.config.imageToAscii === 'failure+success';
+        makeHtmlSnapshot = context.config.htmlOnExpect === 'failure+success';
+        makeDump = context.config.dumpOnExpect === 'failure+success';
       } else {
         protractorUtil.test.failedExpectations.push(expectation);
         makeScreenshotsFromEachBrowsers = context.config.screenshotOnExpect === 'failure+success' || context.config.screenshotOnExpect === 'failure';
         makeAsciiLog = context.config.imageToAscii === 'failure+success' || context.config.imageToAscii === 'failure';
+        makeHtmlSnapshot = context.config.htmlOnExpect === 'failure+success' || context.config.htmlOnExpect === 'failure';
+        makeDump = context.config.dumpOnExpect === 'failure+success' || context.config.dumpOnExpect === 'failure';
       }
     } else {
       console.warn('Calling addExpectationResult before specStarted!');
+    }
+    if (makeDump) {
+      protractorUtil.takeDump(context, function(dump) {
+        expectation.dump = dump;
+        if (context.config.writeReportFreq === 'asap') {
+          protractorUtil.writeReport(context, protractorUtil.newLongRunningOperationCounter());
+        }
+      });
     }
     if (makeScreenshotsFromEachBrowsers) {
       protractorUtil.takeScreenshot(context, function(filename, browserName, finalFile, browserInstance, done) {
@@ -164,6 +217,7 @@ protractorUtil.takeScreenshotOnExpectDone = function(context) {
           browser: browserName,
           when: new Date()
         });
+
         if (makeAsciiLog && !browserInstance.skipImageToAscii) {
           try {
             imageToAscii(finalFile, context.config.imageToAsciiOpts, function(err, converted) {
@@ -186,6 +240,22 @@ protractorUtil.takeScreenshotOnExpectDone = function(context) {
         }
       });
     }
+
+    if (makeHtmlSnapshot) {
+      protractorUtil.takeRawHtml(context, function(filename, browserName, finalFile, browserInstance, done) {
+        expectation.htmls.push({
+          file: filename,
+          browser: browserName,
+          when: new Date()
+        });
+        if (context.config.writeReportFreq === 'asap') {
+          protractorUtil.writeReport(context, done);
+        } else {
+          done();
+        }
+      });
+    }
+
     if (context.config.withLogs) {
       protractorUtil.takeLogs(context, function(logs, browserName, done) {
         expectation.logs.push({
@@ -204,16 +274,32 @@ protractorUtil.takeScreenshotOnExpectDone = function(context) {
 };
 
 
-protractorUtil.takeScreenshotOnSpecDone = function(result, context, test) {
+protractorUtil.takeOnSpecDone = function(result, context, test) {
 
   var makeScreenshotsFromEachBrowsers = false;
+  var makeHtmlSnapshot = false;
+  var makeDump = false;
   if (result.failedExpectations.length === 0) {
     makeScreenshotsFromEachBrowsers = context.config.screenshotOnSpec === 'failure+success';
+    makeHtmlSnapshot = context.config.htmlOnSpec === 'failure+success';
+    makeDump = context.config.dumpOnSpec === 'failure+success';
   } else {
     makeScreenshotsFromEachBrowsers = context.config.screenshotOnSpec === 'failure+success' || context.config.screenshotOnSpec === 'failure';
+    makeHtmlSnapshot = context.config.htmlOnSpec === 'failure+success' || context.config.htmlOnSpec === 'failure';
+    makeDump = context.config.dumpOnSpec === 'failure+success' || context.config.dumpOnSpec === 'failure';
   }
   if (result.status === 'disabled' || result.status === 'pending') {
     makeScreenshotsFromEachBrowsers = false;
+    makeHtmlSnapshot = false;
+    makeDump = false;
+  }
+  if (makeDump) {
+    protractorUtil.takeDump(context, function(dump) {
+      test.specDump = dump;
+      if (context.config.writeReportFreq === 'asap' || context.config.writeReportFreq === 'spec') {
+        protractorUtil.writeReport(context, protractorUtil.newLongRunningOperationCounter());
+      }
+    });
   }
   if (makeScreenshotsFromEachBrowsers) {
     protractorUtil.takeScreenshot(context, function(file, browserName, finalFile, browserInstance, done) {
@@ -228,8 +314,26 @@ protractorUtil.takeScreenshotOnSpecDone = function(result, context, test) {
         done();
       }
     });
+
+
   }
-  if (context.config.withLogs) {
+
+  if (makeHtmlSnapshot) {
+    protractorUtil.takeRawHtml(context, function(file, browserName, finalFile, browserInstance, done) {
+      test.specHtmls.push({
+        file: file,
+        browser: browserName,
+        when: new Date()
+      });
+      if (context.config.writeReportFreq === 'asap' || context.config.writeReportFreq === 'spec') {
+        protractorUtil.writeReport(context, done);
+      } else {
+        done();
+      }
+    });
+  }
+
+  if (context.config.screenshotOnSpec != 'none' && context.config.withLogs) {
     protractorUtil.takeLogs(context, function(logs, browserName, done) {
       test.specLogs.push({
         logs: logs,
@@ -305,7 +409,8 @@ protractorUtil.joinReports = function(context, done) {
       data.stat.pending += report.stat.pending || 0;
       data.stat.disabled += report.stat.disabled || 0;
     } catch (err) {
-      protractorUtil.logError('Unknown error while process report %s', reports[i]);
+      // need to refactor cb to promises + use promise all to avoid concurent writes and reads
+      // protractorUtil.logDebug('Unknown error while process report %s', reports[i]);
       return done(err);
     }
   }
@@ -356,6 +461,7 @@ protractorUtil.registerJasmineReporter = function(context) {
         start: moment(),
         specScreenshots: [],
         specLogs: [],
+        specHtmls: [],
         failedExpectations: [],
         passedExpectations: []
       };
@@ -380,9 +486,7 @@ protractorUtil.registerJasmineReporter = function(context) {
       }
     },
     specDone: function(result) {
-      if (context.config.screenshotOnSpec != 'none') {
-        protractorUtil.takeScreenshotOnSpecDone(result, context, protractorUtil.test); //exec async operation
-      }
+      protractorUtil.takeOnSpecDone(result, context, protractorUtil.test); //exec async operation
 
       //calculate total fails, success and so on
       if (!protractorUtil.stat[result.status]) {
@@ -497,7 +601,7 @@ protractorUtil.failTestOnErrorLog = function(context) {
 };
 
 /**
- * Initialize configurtion
+ * Initialize configuration
  */
 protractorUtil.prototype.setup = function() {
   var defaultSettings = {
@@ -505,13 +609,18 @@ protractorUtil.prototype.setup = function() {
     clearFoldersBeforeTest: true,
     withLogs: true,
     screenshotOnExpect: 'failure+success',
+    htmlOnExpect: 'failure',
+    dumpOnExpect: 'failure',
     verbose: 'info',
     screenshotOnSpec: 'failure+success',
+    htmlOnSpec: 'failure',
+    dumpOnSpec: 'none',
     pauseOn: 'never',
     imageToAscii: 'none',
     imageToAsciiOpts: {
       bg: true
     },
+    dump: null,
     htmlReport: true,
     writeReportFreq: 'end',
     speak: false
@@ -538,6 +647,14 @@ protractorUtil.prototype.setup = function() {
       console.error(err);
     } else {
       protractorUtil.logDebug(self.config.screenshotPath + '/screenshots' + ' folder created!');
+    }
+  });
+
+  mkdirp.sync(this.config.screenshotPath + '/htmls', function(err) {
+    if (err) {
+      console.error(err);
+    } else {
+      protractorUtil.logDebug(self.config.screenshotPath + '/htmls' + ' folder created!');
     }
   });